python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2

# OCR and Computer Vision
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
easyocr==1.7.0

# AI and Machine Learning
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
sentence-transformers==2.2.2

# Vector Database and RAG
chromadb==0.4.18
faiss-cpu==1.7.4
langchain==0.0.350
langchain-community==0.0.1

# LLM Integration
openai==1.3.7
anthropic==0.7.7
groq==0.4.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
